增加了登录功能，并对路由进行了guard

+++ web-src/src/app/blog-module/details-component/details-component.component.ts
@@ -2,18 +2,22 @@ import {Component, OnInit} from "@angular/core";
 import {ActivatedRoute, Router} from "@angular/router";
 import {Article} from "../../../models/article";
 import {AppService} from "../../app.service";
+import {AuthGuard} from "../../auth.guard";
 
 @Component({
   selector: 'app-details-component',
   templateUrl: './details-component.component.html',
   styleUrls: ['./details-component.component.css'],
-  providers: [AppService]
+  providers: [AppService, AuthGuard]
 })
 export class DetailsComponentComponent implements OnInit {
 
   article: Article;
 
-  constructor(private route: ActivatedRoute, private router: Router, private appService: AppService) {
+  constructor(private route: ActivatedRoute,
+              private router: Router,
+              private appService: AppService,
+              private authGuard: AuthGuard) {
   }
 
   ngOnInit() {
@@ -28,12 +32,17 @@ export class DetailsComponentComponent implements OnInit {
   }
 
   onDeleteArticle() {
-    this.appService.deleteArticle(this.article._id).subscribe(res => {
-      if(res.status == 200) {
-        alert("delete successfully")
-        this.router.navigate(["/"]);
-      }
-    })
+    if (this.authGuard.canActivate()==true){
+      this.appService.deleteArticle(this.article._id).subscribe(res => {
+        if (res.status == 200) {
+          alert("delete successfully")
+          this.router.navigate(["/"]);
+        }
+      })
+    } else {
+      this.router.navigate(["/admin/login"])
+    }
+
   }
 
 }
